+++ business/usuario.py
@@ -1,7 +1,6 @@
 import csv
 import re
 
-
 class Usuario:
     def __init__(self, nome: str, data_nascimento: str, cpf: str, endereco: str):
         if not re.match(r"\d{2}/\d{2}/\d{4}", data_nascimento):
@@ -15,6 +14,7 @@ class Usuario:
         self.cpf = cpf
         self.endereco = endereco
         self.contas: list = []
+        
         with open("usuarios.csv", "a", newline="") as file:
             writer = csv.writer(file)
             writer.writerow([self.nome, self.data_nascimento, self.cpf, self.endereco])
