+++ modules/cadastro_usuario.py
@@ -1,0 +1,14 @@
+
+from business.usuario import Usuario
+
+def CadastrarUsuario() -> Usuario:
+
+    nome = input("Digite o nome de usuário: ")
+    data_nascimento = input("Digite a data de nascimento (dd/mm/aaaa): ")
+    cpf = input("Digite o CPF (somente números): ")
+    endereco = input("Digite o endereço (logradouro, nro - bairro - cidade/sigla estado): ")
+    try:
+        usuario = Usuario(nome, data_nascimento, cpf, endereco)
+        print(f"Usuário {nome} criado com sucesso!")
+    except ValueError as e:
+        print(e)
